Clear ShowPatchOutlines when entering map editor
The map editor set Editor.State = EditorSelecting in two situations:
when the user started the editor, and when the user clicked the
SelectButton.  The editor also set Editor.ShowPatchOutlines = false
in the latter case but not in the former.  That variable could thus
be left true if the user clicked PatchButton in the editor, exit
to the main menu, and restarted the editor.