Preserve MapFieldLandUnit etc. in CPatchManager
When copying flags from the patch type to a map field, copy only those
flags that should be controlled by patches, and preserve the ones that
should be controlled by units.

Also, if no patch is found for the map field (i.e. there is a hole in
the map), flag it unpassable.

Fixes: http://savannah.nongnu.org/bugs/?29731
 (Replacing patches in the map editor discards CMapField::Flags
  set because of units)